+++ betty/asset.py
@@ -14,7 +14,8 @@ from typing import TYPE_CHECKING, final, override
 from betty.concurrent import AsynchronizedLock
 from betty.functools import unique
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugin import Plugin, PluginTypeDefinition, Requires
+from betty.plugin import PluginTypeDefinition
+from betty.plugin.cls import Plugin
 from betty.plugin.factory import PluginManufacturer
 from betty.plugin.ordered import Order, OrderedPluginDefinition
 from betty.service.plugin import ServicePluginDefinition
@@ -24,6 +25,7 @@ if TYPE_CHECKING:
     from collections.abc import AsyncIterable, Iterable, Mapping, Sequence
 
     from betty.machine_name import ResolvableMachineName
+    from betty.requirement import Requires
 
 
 class AssetError(Exception):
