+++ betty/entity/__init__.py
@@ -14,7 +14,8 @@ from betty.json.linked_data import (
 from betty.json.schema import JsonSchemaReference, String
 from betty.locale.localizable.gettext import _, ngettext
 from betty.locale.localize import DEFAULT_LOCALIZER
-from betty.plugin import Plugin, PluginDefinition, PluginTypeDefinition
+from betty.plugin import PluginTypeDefinition
+from betty.plugin.cls import Plugin, PluginClsDefinition
 from betty.string import kebab_case_to_lower_camel_case
 
 if TYPE_CHECKING:
@@ -143,7 +144,7 @@ class Entity(LinkedDataDumpableWithSchemaJsonLdObject, Plugin["EntityDefinition"
         "Entities represent the information in your ancestry, such as people and places."
     ),
 )
-class EntityDefinition(CountableHumanFacingDefinition, PluginDefinition[Entity]):
+class EntityDefinition(CountableHumanFacingDefinition, PluginClsDefinition[Entity]):
     """
     .. plugin_type:: entity.
     """
