+++ betty/gender/__init__.py
@@ -8,7 +8,8 @@ from typing import final, override
 
 from betty.definition.human_facing import CountableHumanFacingDefinition
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugin import Plugin, PluginDefinition, PluginTypeDefinition
+from betty.plugin import PluginTypeDefinition
+from betty.plugin.cls import Plugin, PluginClsDefinition
 from betty.plugin.factory import PluginManufacturer
 
 
@@ -25,7 +26,7 @@ class Gender(Plugin["GenderDefinition"]):
     label_plural=_("Genders"),
     label_countable=ngettext("{count} gender", "{count} genders"),
 )
-class GenderDefinition(CountableHumanFacingDefinition, PluginDefinition[Gender]):
+class GenderDefinition(CountableHumanFacingDefinition, PluginClsDefinition[Gender]):
     """
     .. plugin_type:: gender.
