+++ betty/gramps/loader.py
@@ -40,7 +40,8 @@ from betty.locale.localizable.markup import AnyEnumeration
 from betty.locale.localizable.static import StaticTranslations
 from betty.media_type import InvalidMediaType, MediaType
 from betty.place_type import PlaceTypeManufacturer
-from betty.plugin import Plugin, PluginDefinition
+from betty.plugin import PluginDefinition
+from betty.plugin.cls import Plugin
 from betty.plugin.error import PluginNotFound
 from betty.plugins.entity.citation import Citation
 from betty.plugins.entity.enclosure import Enclosure
