+++ betty/link.py
@@ -14,7 +14,9 @@ from betty.locale.localizable import (
     resolve_localizable,
 )
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugin import Plugin, PluginTypeDefinition
+from betty.machine_name import ResolvableMachineName
+from betty.plugin import PluginTypeDefinition
+from betty.plugin.cls import Plugin
 from betty.plugin.factory import PluginManufacturer
 from betty.plugin.ordered import Order, OrderedPluginDefinition
 from betty.service.plugin import ServicePluginDefinition
