+++ betty/plugin/error.py
@@ -9,11 +9,8 @@ from typing import TYPE_CHECKING
 from betty.exception import HumanFacingException
 from betty.locale.localizable.gettext import _
 from betty.locale.localizable.markup import Paragraph, do_you_mean
-from betty.plugin import (
-    PluginDefinition,
-    ResolvablePluginId,
-    resolve_plugin_id,
-)
+from betty.plugin import PluginDefinition
+from betty.plugin.resolve import ResolvablePluginId, resolve_plugin_id
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
