+++ betty/plugin/schema.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, final
 
 from betty.json.schema import Enum
 from betty.locale.localize import DEFAULT_LOCALIZER
-from betty.plugin import Plugin, PluginDefinition, PluginTypeDefinition
+from betty.plugin import PluginDefinition, PluginTypeDefinition
 from betty.string import kebab_case_to_lower_camel_case
 
 if TYPE_CHECKING:
@@ -16,18 +16,15 @@ if TYPE_CHECKING:
 
 
 @final
-class PluginIdSchema[
-    PluginDefinitionT: PluginDefinition = PluginDefinition,
-    PluginT: Plugin = Plugin,
-](Enum):
+class PluginIdSchema[PluginDefinitionT: PluginDefinition = PluginDefinition](Enum):
     """
     The JSON schema for the IDs of the plugins in this repository.
     """
 
     def __init__(
         self,
-        plugin_type: PluginTypeDefinition[PluginT, PluginDefinitionT],
-        plugins: Iterable[PluginDefinition[PluginT]],
+        plugin_type: PluginTypeDefinition[PluginDefinitionT],
+        plugins: Iterable[PluginDefinitionT],
         /,
     ):
         label = plugin_type.label.localize(DEFAULT_LOCALIZER)
