+++ betty/plugins/jinja_test/plugin.py
@@ -7,13 +7,13 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from betty.jinja.test import JinjaTest, JinjaTestDefinition
-from betty.plugin import Plugin as PluginType
-from betty.plugin import PluginDefinition
+from betty.plugin.cls import Plugin as PluginType
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
 
     from betty.machine_name import MachineName
+    from betty.plugin import PluginDefinition
     from betty.plugin.discovery import ResolvableDiscovery
     from betty.service.level import ServiceLevel
