+++ betty/role/__init__.py
@@ -8,7 +8,8 @@ from typing import final, override
 
 from betty.definition.human_facing import CountableHumanFacingDefinition
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugin import Plugin, PluginDefinition, PluginTypeDefinition
+from betty.plugin import PluginTypeDefinition
+from betty.plugin.cls import Plugin, PluginClsDefinition
 from betty.plugin.factory import PluginManufacturer
 
 
@@ -25,7 +26,7 @@ class Role(Plugin["RoleDefinition"]):
     label_plural=_("Roles"),
     label_countable=ngettext("{count} role", "{count} roles"),
 )
-class RoleDefinition(CountableHumanFacingDefinition, PluginDefinition[Role]):
+class RoleDefinition(CountableHumanFacingDefinition, PluginClsDefinition[Role]):
     """
     .. plugin_type:: role.
     """
