+++ betty/test_utils/service/plugin.py
@@ -4,7 +4,8 @@ Test utilities for :py:mod:`betty.service.plugin`.
 
 from typing import final, override
 
-from betty.plugin import Plugin, PluginTypeDefinition
+from betty.plugin import PluginTypeDefinition
+from betty.plugin.cls import Plugin
 from betty.plugin.factory import PluginManufacturer
 from betty.service.plugin import ServicePluginDefinition
 from betty.test_utils.locale.localizable import DUMMY_COUNTABLE_LOCALIZABLE
