+++ betty/tests/plugins/extension/maps/test_ui.py
@@ -2,7 +2,7 @@ from collections.abc import Iterable
 from typing import override
 
 from betty.extension import ExtensionDefinition
-from betty.plugin import ResolvablePluginId
+from betty.plugin.resolve import ResolvablePluginId
 from betty.test_utils.project.extension.maps import MapsTestBase
