+++ betty/tests/plugins/extension/raspberry_mint/test_maps_ui.py
@@ -2,7 +2,7 @@ from collections.abc import Iterable
 from typing import override
 
 from betty.extension import ExtensionDefinition
-from betty.plugin import ResolvablePluginId
+from betty.plugin.resolve import ResolvablePluginId
 from betty.plugins.extension.raspberry_mint import RaspberryMint
 from betty.test_utils.project.extension.maps import MapsTestBase
